+++ src/internal/util/gitHub.ts
@@ -1,0 +1,8 @@
+
+import * as impl from "../../util/gitHub";
+
+/**
+ * Exported only for backward compatible. Use util package.
+ * @type {(token: string, user: string, repo: string, path: string) => Promise<boolean>}
+ */
+export const hasFile = impl.hasFile;