+++ src/project/local/NodeFsLocalProject.ts
@@ -8,7 +8,7 @@ import * as gs from "glob-stream";
 import * as stream from "stream";
 import { deleteFolderRecursive } from "../../internal/util/file";
 import { logger } from "../../internal/util/logger";
-import { RepoRef } from "../../operations/common/RepoId";
+import { RepoRef, SimpleRepoId } from "../../operations/common/RepoId";
 import { AbstractProject } from "../support/AbstractProject";
 import { toPromise } from "../util/projectUtils";
 import { isLocalProject, LocalProject } from "./LocalProject";
@@ -73,11 +73,11 @@ export class NodeFsLocalProject extends AbstractProject implements LocalProject 
     /**
      * Note: this does not validate existence of the target
      * directory, so using it except in tests should be avoided
-     * @param {RepoRef} id
+     * @param {RepoRef} ident identification of the repo
      * @param {string} baseDir
      */
-    public constructor(id: RepoRef, baseDir: string) {
-        super(id);
+    public constructor(ident: RepoRef | string, baseDir: string) {
+        super(typeof ident === "string" ? new SimpleRepoId(undefined, ident) : ident);
         // TODO not sure why app-root-path can return something weird and this coercion is necessary
         this.baseDir = "" + baseDir;
     }