+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/automation-client-ts/compare/0.2.4...HEAD
 
+## [0.2.5][] - 2017-10-26
+
+[0.2.5]: https://github.com/atomist/automation-client-ts/compare/0.2.4...0.2.5
+
+Backward compatibility release
+
+### Fixed
+
+-   Much closer to backward compatible with 0.2.3 project operations than 0.2.4.
+
 ## [0.2.4][] - 2017-10-26
 
 [0.2.4]: https://github.com/atomist/automation-client-ts/compare/0.2.3...0.2.4